+++ src/dync/model/Code_Repository.java
@@ -11,7 +11,7 @@ public class Code_Repository {
 	public static final String CODE_REPOSITORY	= "CODE_REPOSITORY";
 	public static final String USER_ID			= "USER_ID";
 
-	private int code_repository;
+	private long code_repository;
 	private int user_id;
 
 	public Code_Repository() {
@@ -24,11 +24,11 @@ public class Code_Repository {
 		this.user_id = user_id;
 	}
 
-	public int getCode_repository() {
+	public long getCode_repository() {
 		return code_repository;
 	}
 
-	public void setCode_repository(int code_repository) {
+	public void setCode_repository(long code_repository) {
 		this.code_repository = code_repository;
 	}
